Moved stuff to local. Moved some aliases to mac only

+++ .bash_aliases
@@ -1,7 +1,7 @@
 #aliases
 
 #Ensure that ls always shows color
-if [[ "$OSTYPE" =~ ^darwin ]]; then
+if [[ "$OSTYPE" == darwin* ]]; then
 	alias ls="command ls -GAsF"
 else
 	alias ls=" command ls --color -AsF"
@@ -21,6 +21,7 @@ alias javau='java -cp .:/Users/adno4/bin/ant-1.8/junit-4.10.jar org.junit.runner
 alias javaps='java -agentlib:hprof=cpu=samples'
 alias javapt='java -agentlib:hprof=cpu=times'
 alias fn='find . -name'
+# Allow the use of sudo with aliases.
 alias sudo='sudo '
 alias path='echo -e ${PATH//:/\\n}'
 #My habbit is to just type "vi"
@@ -34,19 +35,17 @@ alias ips="ifconfig -a | grep -o 'inet6\? \(\([0-9]\+\.[0-9]\+\.[0-9]\+\.[0-9]\+
 # Enhanced WHOIS lookups
 alias whois="whois -h whois-servers.net"
 
+# Do these things if this is a mac.
+if [[ "$OSTYPE" == darwin* ]]; then
 # Empty the Trash on all mounted volumes and the main HDD
-# Also, clear Apple’s System Logs to improve shell startup speed
-alias emptytrash="sudo rm -rfv /Volumes/*/.Trashes; sudo rm -rfv ~/.Trash; sudo rm -rfv /private/var/log/asl/*.asl"
-alias dstore-clean='find . -type f -name .DS_Store -print0 | xargs -0 rm'
+  alias emptytrash="sudo rm -rfv /Volumes/*/.Trashes; sudo rm -rfv ~/.Trash; sudo rm -rfv /private/var/log/asl/*.asl"
 
 # Hide/show all desktop icons (useful when presenting)
-alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
-alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
+  alias hidedesktop="defaults write com.apple.finder CreateDesktop -bool false && killall Finder"
+  alias showdesktop="defaults write com.apple.finder CreateDesktop -bool true && killall Finder"
 
 # Stuff I never really use but cannot delete either because of http://xkcd.com/530/
-alias stfu="osascript -e 'set volume output muted true'"
-alias pumpitup="osascript -e 'set volume 7'"
-alias hax="growlnotify -a 'Activity Monitor' 'System error' -m 'WTF R U DOIN'"
-
-alias restartService="service memcached restart && service mysql restart && service apache2 restart"
-alias drb="drush -l bz.bz"
+  alias stfu="osascript -e 'set volume output muted true'"
+  alias pumpitup="osascript -e 'set volume 7'"
+  alias hax="growlnotify -a 'Activity Monitor' 'System error' -m 'WTF R U DOIN'"
+fi
